The AbstractTopDownSet class

Abstract/VIRTUAL parent class for TopDownSet and NCBSet to provide common interface.

# S4 method for AbstractTopDownSet,ANY,ANY,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for AbstractTopDownSet,ANY,missing
[[(x, i, j, ...)

# S4 method for AbstractTopDownSet,ANY,missing
[[(x, i, j, ...) <- value

# S4 method for AbstractTopDownSet
$(x, name)

# S4 method for AbstractTopDownSet
$(x, name) <- value

# S4 method for AbstractTopDownSet
assayData(object)

# S4 method for AbstractTopDownSet
colData(object)

# S4 method for AbstractTopDownSet
colData(object, ...) <- value

# S4 method for AbstractTopDownSet
conditionData(object, ...)

# S4 method for AbstractTopDownSet
conditionData(object, ...) <- value

# S4 method for AbstractTopDownSet
dim(x)

# S4 method for AbstractTopDownSet
dimnames(x)

# S4 method for AbstractTopDownSet
removeEmptyConditions(object)

# S4 method for AbstractTopDownSet
rowViews(object, ...)

# S4 method for AbstractTopDownSet
show(object)

# S4 method for AbstractTopDownSet
summary(object, what = c("rows", "columns"),
  ...)

Arguments

i, j
numeric, logical or character, indices specifying elements to extract or replace.

drop
logical, currently ignored.

value
replacment value.

name
character name of an (non)existing column in colData.

object, x
AbstractTopDownSet

what
character, specifies whether "rows" or "columns" should be summarized.

…
arguments passed to internal/other methods.

Details

This class just provides a common interface. It is not intended for direct use by the user. Please see TopDownSet for an example usage of its child class.

Methods (by generic)

[: Subset operator.

For i numeric, logical or character vectors or empty (missing) or NULL are supported. Subsetting is done on the fragment/bond (row) level. character indices could be names (e.g. c("a1", "b1", "c1", "c2", "c3")) or types (e.g. c("c", "x")) of the fragments for TopDownSet objects, or names of the bonds (e.g. c("bond001")) for NCBSet objects. j could be a numeric or logical vector and subsetting is done on the condition/run (column) level.

[[: Subset operator. i could be a numeric or logical vector and subsetting is done on the condition/run (column) level.

[[<-: Setter for a column in the colData slot.

The [[&lt;- operator is used to add/replace a single column of the colData DataFrame.

$: Accessor for columns in the colData slot.

The $ simplifies the accession of a single column of the colData. It is identical to the [[ operator.

$<-: Setter for a column in the colData slot.

The $&lt;- operator is used to add/replace a single column of the colData DataFrame. It is identical to the [[&lt;- operator.

assayData: Accessor for the assay slot.

Returns a Matrix::dgCMatrix that stores the intensity/coverage information of AbstractTopDownSet object.

colData: Accessor for the colData slot.

Returns a S4Vectors::DataFrame that stores metadata for the conditons/runs (columns) of the AbstractTopDownSet object.

colData<-: Setter for the colData slot.

Replaces metadata for the conditons/runs (columns) of the AbstractTopDownSet object.

conditionData: Accessor for the colData slot.

An alias for colData.

conditionData<-: Setter for the colData slot.

An alias for colData&lt;-.

dim: Accessor for dimensions.

Returns a numeric with number of fragments/bonds (rows) and conditions/runs (columns).

dimnames: Accessor for dimension names.

Returns a list with names for the fragments/bonds (rows) and for the conditions/runs (columns).

removeEmptyConditions: Remove empty conditions/runs.

Removes conditions/runs (columns) without any intensity/coverage information from the AbstractTopDownSet object. It returns a modified AbstractTopDownSet object.

rowViews: Accessor for the rowViews slot.

Depending on the implementation it returns an FragmentViews object for TopDownSet objects or an Biostrings::XStringViews object for NCBSet objects.

summary: Summary statistics.

Returns a matrix with some statistics: number of fragments, total/min/first quartile/median/mean/third quartile/maximum of intensity values.

Slots

rowViews

Biostrings::XStringViews, information about fragments/bonds (name, type, sequence, mass, charge), see Biostrings::XStringViews and FragmentViews for details.

colData

S4Vectors::DataFrame, information about the MS2 experiments and the fragmentation conditions.

assay

Matrix::dgCMatrix, intensity/coverage values of the fragments/bonds. The rows correspond to the fragments/bonds and the columns to the condition/run. It just stores values that are different from zero.

files

character, files that were imported.

tolerance

double, tolerance in ppm that were used for matching the experimental m/z values to the theoretical fragments.

processing

character, log messages.

See also

TopDownSet and NCBSet which both implement/use this interface. These manual pages also provide some example code.

FragmentViews (and Biostrings::XStringViews) for the row view interface.

Matrix::dgCMatrix for technical details about the intensity/coverage storage.