The NCBSet class

The NCBSet class is a container for a top-down proteomics experiment similar to the TopDownSet but instead of intensity values it just stores the information if a bond is covered by a N-terminal (encoded as 1), a C-terminal (encoded as 2) and/or both fragments (encoded as 3).

# S4 method for NCBSet
bestConditions(object, n = ncol(object), minN = 0L, ...)

# S4 method for NCBSet
fragmentationMap(object, nCombinations = 10,
  cumCoverage = TRUE, labels = colnames(object), ...)

# S4 method for NCBSet
show(object)

# S4 method for NCBSet
summary(object, what = c("conditions", "bonds"), ...)

Arguments

object
NCBSet

n
integer, max number of combinations/iterations.

minN
integer stop if there are less than minN additional fragments

nCombinations
integer, number of combinations to show (0 to avoid plotting them at all).

cumCoverage
logical, if TRUE (default) cumulative coverage of combinations is shown.

labels
character, overwrite x-axis labels.

what
character, specifies whether "conditions" (columns; default) or "bonds" (rows) should be summarized.

…
arguments passed to internal/other methods. added. newly added fragments third column: number of newly covered bonds.

Methods (by generic)

bestConditions: Best combination of conditions.

Finds the best combination of conditions for highest coverage of bonds. Use n to limit the number of iterations and combinations that should be returned. If minN is set at least minN fragments have to be added to the combinations. The function returns a 2-column matrix. The first column contains the index of the condition (column number) and the second one the newly added number of fragments.

fragmentationMap: Plot fragmentation map.

Plots a fragmentation map of the Protein. Use nCombinations to add another plot with nCombinations combined conditions. If cumCoverage is TRUE (default) these combinations increase the coverage cumulatively.

summary: Summary statistics.

Returns a matrix with some statistics: number of fragments, total/min/first quartile/median/mean/third quartile/maximum of intensity values.

Slots

rowViews

Biostrings::XStringViews, information about bonds (name, start, end, width, sequence), see Biostrings::XStringViews for details.

colData

S4Vectors::DataFrame, information about the MS2 experiments and the fragmentation conditions.

assay

Matrix::dgCMatrix, coverage values of the bonds. The rows correspond to the bonds and the columns to the condition/run. It just stores values that are different from zero. If a bond is covered by an N-terminal fragment its encoded with 1, by an C-terminal fragmentl with 2 and by both fragment types by 3 respectively.

files

character, files that were imported.

tolerance

double, tolerance in ppm that were used for matching the experimental mz values to the theoretical fragments.

processing

character, log messages.

See also

An NCBSet is generated from an TopDownSet object.

Biostrings::XStringViews for the row view interface.

Matrix::dgCMatrix for technical details about the coverage storage.