Expand MS Conditions

Create a data.frame of all possible combinations of the given arguments. It ensures that just arguments are applied that yield a valid MethodModification.xml file.

expandMs1Conditions(..., family = "Calcium", version = "3.2")

expandTms2Conditions(
  ActivationType = c("CID", "HCD", "ETD", "UVPD"),
  ...,
  MassList = NULL,
  family = "Calcium",
  version = "3.2"
)

Arguments

...
further named arguments, used to create the combination of conditions.

family
character, currently just Calcium is supported

version
character, currently 3.1, 3.2 (default), 3.3 are supported

ActivationType
character, ActivationType for TMS2, either CID, HCD, ETD, or UVPD.

MassList
matrix, 2 columns (mass, z) for targeted mass list, or NULL (default) to not overwrite targeted mass.

Value

data.frame with all possible combinations of conditions/settings.

See also

Examples

expandMs1Conditions(FirstMass=100, LastMass=400)

#>   FirstMass LastMass
#> 1       100      400

expandTms2Conditions(
     ActivationType="CID",
     OrbitrapResolution="R120K",
     IsolationWindow=1,
     MaxITTimeInMS=200,
     Microscans=as.integer(40),
     AgcTarget=c(1e5, 5e5, 1e6),
     CIDCollisionEnergy=c(NA, seq(7, 35, 7)),
     MassList=cbind(mz=c(560.6, 700.5, 933.7), z=rep(1, 3))
)

#>                   MassList ActivationType OrbitrapResolution IsolationWindow
#> 1  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 2  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 3  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 4  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 5  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 6  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 7  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 8  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 9  560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 10 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 11 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 12 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 13 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 14 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 15 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 16 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 17 560.6/1 700.5/1 933.7/1            CID              R120K               1
#> 18 560.6/1 700.5/1 933.7/1            CID              R120K               1
#>    MaxITTimeInMS Microscans AgcTarget CIDCollisionEnergy
#> 1            200         40     1e+05                 NA
#> 2            200         40     5e+05                 NA
#> 3            200         40     1e+06                 NA
#> 4            200         40     1e+05                  7
#> 5            200         40     5e+05                  7
#> 6            200         40     1e+06                  7
#> 7            200         40     1e+05                 14
#> 8            200         40     5e+05                 14
#> 9            200         40     1e+06                 14
#> 10           200         40     1e+05                 21
#> 11           200         40     5e+05                 21
#> 12           200         40     1e+06                 21
#> 13           200         40     1e+05                 28
#> 14           200         40     5e+05                 28
#> 15           200         40     1e+06                 28
#> 16           200         40     1e+05                 35
#> 17           200         40     5e+05                 35
#> 18           200         40     1e+06                 35