The AbstractTopDownSet class

Abstract/VIRTUAL parent class for TopDownSet and NCBSet to provide common interface.

# S4 method for AbstractTopDownSet,ANY,ANY,ANY
[(x, i, j, ..., drop = FALSE)

# S4 method for AbstractTopDownSet,ANY,missing
[[(x, i, j, ...)

# S4 method for AbstractTopDownSet,ANY,missing
[[(x, i, j, ...) <- value

# S4 method for AbstractTopDownSet
$(x, name)

# S4 method for AbstractTopDownSet
$(x, name) <- value

# S4 method for AbstractTopDownSet
assayData(object)

# S4 method for AbstractTopDownSet
colData(object)

# S4 method for AbstractTopDownSet
colData(object, ...) <- value

# S4 method for AbstractTopDownSet,AbstractTopDownSet
combine(x, y)

# S4 method for AbstractTopDownSet
conditionData(object, ...)

# S4 method for AbstractTopDownSet
conditionData(object, ...) <- value

# S4 method for AbstractTopDownSet
conditionNames(object)

# S4 method for AbstractTopDownSet
dim(x)

# S4 method for AbstractTopDownSet
dimnames(x)

# S4 method for AbstractTopDownSet
removeEmptyConditions(object)

# S4 method for AbstractTopDownSet
rowViews(object, ...)

# S4 method for AbstractTopDownSet
show(object)

# S4 method for AbstractTopDownSet
summary(object, what = c("rows", "columns"),
  ...)

# S4 method for AbstractTopDownSet
updateConditionNames(object,
  sampleColumns = c("Mz", "AgcTarget", "EtdReagentTarget", "EtdActivation",
  "CidActivation", "HcdActivation", "UvpdActivation"),
  verbose = interactive())

# S4 method for TopDownSet
updateMedianInjectionTime(object, by = list(Mz =
  object$Mz, AgcTarget = object$AgcTarget))

Arguments

i, j
numeric, logical or character, indices specifying elements to extract or replace.

drop
logical, currently ignored.

value
replacment value.

name
character name of an (non)existing column in colData.

object, x
AbstractTopDownSet

y
AbstractTopDownSet

what
character, specifies whether "rows" or "columns" should be summarized.

sampleColumns
character, column names of the colData() used to define a sample (technical replicate). This is used to add the Sample column (used for easier aggregation, etc.).

verbose
logical, verbose output?

by
list, grouping information.

…
arguments passed to internal/other methods.

Value

This is an Abstract/VIRTUAL class to provide a common interface for TopDownSet and NCBSet. It is not possible to create an AbstractTopDownSet object.

Details

This class just provides a common interface. It is not intended for direct use by the user. Please see TopDownSet for an example usage of its child class.

Methods (by generic)

[: Subset operator.

For i numeric, logical or character vectors or empty (missing) or NULL are supported. Subsetting is done on the fragment/bond (row) level. character indices could be names (e.g. c("a1", "b1", "c1", "c2", "c3")) or types (e.g. c("c", "x")) of the fragments for TopDownSet objects, or names of the bonds (e.g. c("bond001")) for NCBSet objects. j could be a numeric or logical vector and subsetting is done on the condition/run (column) level.

[[: Subset operator. i could be a numeric or logical vector and subsetting is done on the condition/run (column) level.

[[<-: Setter for a column in the colData slot.

The [[&lt;- operator is used to add/replace a single column of the colData DataFrame.

$: Accessor for columns in the colData slot.

The $ simplifies the accession of a single column of the colData. It is identical to the [[ operator.

$<-: Setter for a column in the colData slot.

The $&lt;- operator is used to add/replace a single column of the colData DataFrame. It is identical to the [[&lt;- operator.

assayData: Accessor for the assay slot.

Returns a Matrix::dgCMatrix that stores the intensity/coverage information of AbstractTopDownSet object.

colData: Accessor for the colData slot.

Returns a S4Vectors::DataFrame that stores metadata for the conditons/runs (columns) of the AbstractTopDownSet object.

colData<-: Setter for the colData slot.

Replaces metadata for the conditons/runs (columns) of the AbstractTopDownSet object.

combine: Combine AbstractTopDownSet objects.

If the rowViews are identically combine allows to combine two or more AbstractTopDownSet objects. Please note that it uses the default sampleColumns to define technical replicates (see readTopDownFiles()).and the default by argument to group ion injection times for the calculation of the median time (see updateMedianInjectionTime()). Both could be modified after combine by calling updateConditionNames() (with modified sampleColumns argument) and updateMedianInjectionTime() (with modified by argument).

conditionData: Accessor for the colData slot.

An alias for colData.

conditionData<-: Setter for the colData slot.

An alias for colData&lt;-.

conditionNames: Accessor for condition names.

Returns a character with names for the conditions/runs (columns).

dim: Accessor for dimensions.

Returns a numeric with number of fragments/bonds (rows) and conditions/runs (columns).

dimnames: Accessor for dimension names.

Returns a list with names for the fragments/bonds (rows) and for the conditions/runs (columns).

removeEmptyConditions: Remove empty conditions/runs.

Removes conditions/runs (columns) without any intensity/coverage information from the AbstractTopDownSet object. It returns a modified AbstractTopDownSet object.

rowViews: Accessor for the rowViews slot.

Depending on the implementation it returns an FragmentViews object for TopDownSet objects or an Biostrings::XStringViews object for NCBSet objects.

summary: Summary statistics.

Returns a matrix with some statistics: number of fragments, total/min/first quartile/median/mean/third quartile/maximum of intensity values.

updateConditionNames: Update condition names.

Updates condition names based on sampleColumns from conditionData/colData. Columns with just identical entries are ignored. This method will create/update the colData(object)$Sample column that identifies technical replicates and could be used in other methods.

updateMedianInjectionTime: Update median ion injection times.

Recalculates median ion injection times by a user given grouping variable (default: Mz, AgcTarget). This is useful if you acquire new data and the ion injection time differs across the runs. Use the by argument to provide a list/data.frame of grouping variables, e.g. by=colData(object)[, c("Mz", "AgcTarget", "File")].

Slots

rowViews

Biostrings::XStringViews, information about fragments/bonds (name, type, sequence, mass, charge), see Biostrings::XStringViews and FragmentViews for details.

colData

S4Vectors::DataFrame, information about the MS2 experiments and the fragmentation conditions.

assay

Matrix::dgCMatrix, intensity/coverage values of the fragments/bonds. The rows correspond to the fragments/bonds and the columns to the condition/run. It just stores values that are different from zero.

files

character, files that were imported.

tolerance

double, tolerance in ppm that were used for matching the experimental m/z values to the theoretical fragments.

processing

character, log messages.

See also

TopDownSet and NCBSet which both implement/use this interface. These manual pages also provide some example code.

FragmentViews (and Biostrings::XStringViews) for the row view interface.

Matrix::dgCMatrix for technical details about the intensity/coverage storage.

Examples

## Because AbstractTopDownSet is a VIRTUAL class we could not create any
## object of it. Here we demonstrate the usage with an TopDownSet that
## implements the AbstractTopDownSet interface. See `?"TopDownSet-class"` for
## more details an further examples.

## Example data
data(tds, package="topdownr")

tds

#> TopDownSet object (0.73 Mb)
#> - - - Protein data - - -
#> Amino acid sequence (153): GLSDGEWQQVLNVWGKVEADIAGHGQ...GAMTKALELFRNDIAAKYKELGFQG 
#> Mass : 16964.96
#> Modifications (3): Carbamidomethyl, Acetyl, Met-loss
#> - - - Fragment data - - -
#> Number of theoretical fragments: 1949 
#> Theoretical fragment types (12): a, a*, b, b*, c, ..., x*, y, y*, z, z*
#> Theoretical mass range: [72.04;16925.98]
#> - - - Condition data - - -
#> Number of conditions: 222 
#> Number of scans: 351 
#> Condition variables (57): Scan, PeaksCount, ..., Sample, MedianIonInjectionTimeMs
#> - - - Intensity data - - -
#> Size of array: 1949x351 (2.18% != 0)
#> Number of matched fragments: 14901 
#> Intensity range: [126.42;1454253.25]
#> - - - Processing information - - -
#> [2017-12-28 16:44:52] 14901 fragments [1949;351] matched (tolerance: 25 ppm).
#> [2017-12-28 16:44:52] Condition names updated based on: Mz, AgcTarget, EtdReagentTarget, EtdActivation, CidActivation, HcdActivation. Order of conditions changed. 222 conditions.
#> [2017-12-28 16:44:52] Recalculate median injection time based on: Mz, AgcTarget. 

head(summary(tds))

#>                                Fragments   Total     Min       Q1   Median
#> C1.0e+05_1.0e+06_02.50_07_00_1        51 3429794 5289.57 14133.81 30823.25
#> C1.0e+05_1.0e+06_02.50_07_00_2        49 3470773 3147.54 18107.74 31476.16
#> C1.0e+05_1.0e+06_02.50_14_00_1        49 3242743 3842.46 14155.42 25640.48
#> C1.0e+05_1.0e+06_02.50_14_00_2        49 3329635 3848.07 10651.31 32818.36
#> C1.0e+05_1.0e+06_02.50_21_00          57 3590927 3787.77 12579.74 27829.12
#> C1.0e+05_1.0e+06_02.50_28_00_1        59 3300209 1843.09  9754.08 21028.82
#>                                    Mean       Q3      Max
#> C1.0e+05_1.0e+06_02.50_07_00_1 1759.771 55583.93 607970.2
#> C1.0e+05_1.0e+06_02.50_07_00_2 1780.797 70941.52 641441.3
#> C1.0e+05_1.0e+06_02.50_14_00_1 1663.798 62368.79 669750.9
#> C1.0e+05_1.0e+06_02.50_14_00_2 1708.381 52343.87 601565.8
#> C1.0e+05_1.0e+06_02.50_21_00   1842.446 51986.04 607209.7
#> C1.0e+05_1.0e+06_02.50_28_00_1 1693.283 48242.05 589361.6

# Accessing slots
rowViews(tds)

#> FragmentViews on a 153-letter sequence:
#>   GLSDGEWQQVLNVWGKVEADIAGHGQEVLIRLFTGHPE...SKHPGDFGADAQGAMTKALELFRNDIAAKYKELGFQG
#> Mass:
#>   16964.964625
#> Modifications:
#>   Carbamidomethyl
#>   Acetyl
#>   Met-loss
#> Views:
#>        start end width     mass name  type   z                                  
#>    [1]     1   1     1    72.04 a1    a      1 [G]                              
#>    [2]     1   1     1   100.04 b1    b      1 [G]                              
#>    [3]     1   1     1   101.02 z1    z      1 [G]                              
#>    [4]     1   1     1   102.03 zpH1  z      1 [G]                              
#>    [5]     1   1     1   117.07 c1    c      1 [G]                              
#>    ...   ... ...   ...      ... ...   ...  ... ...                              
#> [1945]     2 153   152 16893.90 x152* x*     1 [LSDGEWQQVLNVWG...NDIAAKYKELGFQG]
#> [1946]     1 152   152 16908.95 b152  b      1 [GLSDGEWQQVLNVW...RNDIAAKYKELGFQ]
#> [1947]     1 152   152 16908.95 c152* c*     1 [GLSDGEWQQVLNVW...RNDIAAKYKELGFQ]
#> [1948]     2 153   152 16910.93 x152  x      1 [LSDGEWQQVLNVWG...NDIAAKYKELGFQG]
#> [1949]     1 152   152 16925.98 c152  c      1 [GLSDGEWQQVLNVW...RNDIAAKYKELGFQ]

colData(tds)

#> DataFrame with 351 rows and 57 columns
#>                                     Scan PeaksCount TotIonCurrent RetentionTime
#>                                <integer>  <integer>     <numeric>     <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1        33        135      13727530      371.4657
#> C1.0e+05_1.0e+06_02.50_07_00_2        34        133      14978619      389.4116
#> C1.0e+05_1.0e+06_02.50_14_00_1        36        124      15124274      425.3971
#> C1.0e+05_1.0e+06_02.50_14_00_2        37        144      14001930      443.2559
#> C1.0e+05_1.0e+06_02.50_21_00          40        148      14008273      497.2891
#> ...                                  ...        ...           ...           ...
#> C1.0e+06_0.0e+00_00.00_00_21_2       265        471       3388050      4764.555
#> C1.0e+06_0.0e+00_00.00_00_28_1       268        233       2060326      4814.021
#> C1.0e+06_0.0e+00_00.00_00_28_2       269        172       1970904      4830.140
#> C1.0e+06_0.0e+00_00.00_00_35_1       271        166       2918202      4863.092
#> C1.0e+06_0.0e+00_00.00_00_35_2       272        231       2861092      4878.678
#>                                BasePeakMz BasePeakIntensity.SpectraInformation
#>                                 <numeric>                            <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1   16942.98                              4631841
#> C1.0e+05_1.0e+06_02.50_07_00_2   16943.00                              5103187
#> C1.0e+05_1.0e+06_02.50_14_00_1   16943.04                              5346571
#> C1.0e+05_1.0e+06_02.50_14_00_2   16942.99                              4668278
#> C1.0e+05_1.0e+06_02.50_21_00     16942.98                              4605146
#> ...                                   ...                                  ...
#> C1.0e+06_0.0e+00_00.00_00_21_2   231.1233                             56502.66
#> C1.0e+06_0.0e+00_00.00_00_28_1   231.1232                             96439.46
#> C1.0e+06_0.0e+00_00.00_00_28_2   231.1231                            104296.80
#> C1.0e+06_0.0e+00_00.00_00_35_1   231.1232                            261401.52
#> C1.0e+06_0.0e+00_00.00_00_35_2   231.1232                            238047.05
#>                                    LowMz    HighMz PrecursorMz
#>                                <numeric> <numeric>   <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1  762.3914  17146.08    1211.701
#> C1.0e+05_1.0e+06_02.50_07_00_2  762.3905  17146.08    1211.701
#> C1.0e+05_1.0e+06_02.50_14_00_1  762.3910  17146.09    1211.701
#> C1.0e+05_1.0e+06_02.50_14_00_2  762.3913  17146.07    1211.701
#> C1.0e+05_1.0e+06_02.50_21_00    762.3912  17146.11    1211.702
#> ...                                  ...       ...         ...
#> C1.0e+06_0.0e+00_00.00_00_21_2  200.1387 5970.0875    1211.716
#> C1.0e+06_0.0e+00_00.00_00_28_1  200.1386 1460.8074    1211.716
#> C1.0e+06_0.0e+00_00.00_00_28_2  200.1386 1351.7438    1211.716
#> C1.0e+06_0.0e+00_00.00_00_35_1  200.1386  959.4907    1211.716
#> C1.0e+06_0.0e+00_00.00_00_35_2  200.1387 1138.6675    1211.716
#>                                                                  SpectrumId
#>                                                                 <character>
#> C1.0e+05_1.0e+06_02.50_07_00_1  controllerType=0 controllerNumber=1 scan=33
#> C1.0e+05_1.0e+06_02.50_07_00_2  controllerType=0 controllerNumber=1 scan=34
#> C1.0e+05_1.0e+06_02.50_14_00_1  controllerType=0 controllerNumber=1 scan=36
#> C1.0e+05_1.0e+06_02.50_14_00_2  controllerType=0 controllerNumber=1 scan=37
#> C1.0e+05_1.0e+06_02.50_21_00    controllerType=0 controllerNumber=1 scan=40
#> ...                                                                     ...
#> C1.0e+06_0.0e+00_00.00_00_21_2 controllerType=0 controllerNumber=1 scan=265
#> C1.0e+06_0.0e+00_00.00_00_28_1 controllerType=0 controllerNumber=1 scan=268
#> C1.0e+06_0.0e+00_00.00_00_28_2 controllerType=0 controllerNumber=1 scan=269
#> C1.0e+06_0.0e+00_00.00_00_35_1 controllerType=0 controllerNumber=1 scan=271
#> C1.0e+06_0.0e+00_00.00_00_35_2 controllerType=0 controllerNumber=1 scan=272
#>                                Condition
#>                                <integer>
#> C1.0e+05_1.0e+06_02.50_07_00_1        12
#> C1.0e+05_1.0e+06_02.50_07_00_2        12
#> C1.0e+05_1.0e+06_02.50_14_00_1        14
#> C1.0e+05_1.0e+06_02.50_14_00_2        14
#> C1.0e+05_1.0e+06_02.50_21_00          16
#> ...                                  ...
#> C1.0e+06_0.0e+00_00.00_00_21_2       158
#> C1.0e+06_0.0e+00_00.00_00_28_1       160
#> C1.0e+06_0.0e+00_00.00_00_28_2       160
#> C1.0e+06_0.0e+00_00.00_00_35_1       162
#> C1.0e+06_0.0e+00_00.00_00_35_2       162
#>                                                                          FilterString.HeaderInformation
#>                                                                                             <character>
#> C1.0e+05_1.0e+06_02.50_07_00_1  FTMS + p NSI sa Full ms2 1211.7012@etd2.50@cid7.00 [100.0000-2000.0000]
#> C1.0e+05_1.0e+06_02.50_07_00_2  FTMS + p NSI sa Full ms2 1211.7012@etd2.50@cid7.00 [100.0000-2000.0000]
#> C1.0e+05_1.0e+06_02.50_14_00_1 FTMS + p NSI sa Full ms2 1211.7014@etd2.50@cid14.00 [100.0000-2000.0000]
#> C1.0e+05_1.0e+06_02.50_14_00_2 FTMS + p NSI sa Full ms2 1211.7014@etd2.50@cid14.00 [100.0000-2000.0000]
#> C1.0e+05_1.0e+06_02.50_21_00   FTMS + p NSI sa Full ms2 1211.7016@etd2.50@cid21.00 [100.0000-2000.0000]
#> ...                                                                                                 ...
#> C1.0e+06_0.0e+00_00.00_00_21_2            FTMS + p NSI Full ms2 1211.7158@hcd21.00 [100.0000-2000.0000]
#> C1.0e+06_0.0e+00_00.00_00_28_1            FTMS + p NSI Full ms2 1211.7160@hcd28.00 [100.0000-2000.0000]
#> C1.0e+06_0.0e+00_00.00_00_28_2            FTMS + p NSI Full ms2 1211.7160@hcd28.00 [100.0000-2000.0000]
#> C1.0e+06_0.0e+00_00.00_00_35_1            FTMS + p NSI Full ms2 1211.7162@hcd35.00 [100.0000-2000.0000]
#> C1.0e+06_0.0e+00_00.00_00_35_2            FTMS + p NSI Full ms2 1211.7162@hcd35.00 [100.0000-2000.0000]
#>                                SupplementalActivation     RtMin       Tic
#>                                                 <Rle> <numeric> <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1                   TRUE   6.19109 159451792
#> C1.0e+05_1.0e+06_02.50_07_00_2                   TRUE   6.49019 171593248
#> C1.0e+05_1.0e+06_02.50_14_00_1                   TRUE   7.08995 178472720
#> C1.0e+05_1.0e+06_02.50_14_00_2                   TRUE   7.38760 159089344
#> C1.0e+05_1.0e+06_02.50_21_00                     TRUE   8.28815 159144752
#> ...                                               ...       ...       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                  FALSE  79.40925   6983628
#> C1.0e+06_0.0e+00_00.00_00_28_1                  FALSE  80.23369   4512664
#> C1.0e+06_0.0e+00_00.00_00_28_2                  FALSE  80.50233   4734344
#> C1.0e+06_0.0e+00_00.00_00_35_1                  FALSE  81.05153   6846192
#> C1.0e+06_0.0e+00_00.00_00_35_2                  FALSE  81.31130   6351402
#>                                BasePeakPosition
#>                                       <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1         1304.930
#> C1.0e+05_1.0e+06_02.50_07_00_2         1304.930
#> C1.0e+05_1.0e+06_02.50_14_00_1         1304.929
#> C1.0e+05_1.0e+06_02.50_14_00_2         1304.930
#> C1.0e+05_1.0e+06_02.50_21_00           1304.930
#> ...                                         ...
#> C1.0e+06_0.0e+00_00.00_00_21_2         231.1233
#> C1.0e+06_0.0e+00_00.00_00_28_1         231.1232
#> C1.0e+06_0.0e+00_00.00_00_28_2         159.0910
#> C1.0e+06_0.0e+00_00.00_00_35_1         231.1232
#> C1.0e+06_0.0e+00_00.00_00_35_2         231.1232
#>                                BasePeakIntensity.HeaderInformation
#>                                                          <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1                             3762917
#> C1.0e+05_1.0e+06_02.50_07_00_2                             3784709
#> C1.0e+05_1.0e+06_02.50_14_00_1                             3987072
#> C1.0e+05_1.0e+06_02.50_14_00_2                             3722252
#> C1.0e+05_1.0e+06_02.50_21_00                               3627429
#> ...                                                            ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                             50586.9
#> C1.0e+06_0.0e+00_00.00_00_28_1                             86218.9
#> C1.0e+06_0.0e+00_00.00_00_28_2                            113096.1
#> C1.0e+06_0.0e+00_00.00_00_35_1                            232769.3
#> C1.0e+06_0.0e+00_00.00_00_35_2                            212045.9
#>                                IonInjectionTimeMs ElapsedScanTimeSec HcdEnergy
#>                                         <numeric>          <numeric>     <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1              5.594             17.883        NA
#> C1.0e+05_1.0e+06_02.50_07_00_2              3.844             17.817        NA
#> C1.0e+05_1.0e+06_02.50_14_00_1              3.345             17.796        NA
#> C1.0e+05_1.0e+06_02.50_14_00_2              5.548             17.885        NA
#> C1.0e+05_1.0e+06_02.50_21_00                5.622             17.887        NA
#> ...                                           ...                ...       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2             33.946             15.801        21
#> C1.0e+06_0.0e+00_00.00_00_28_1             40.343             16.056        28
#> C1.0e+06_0.0e+00_00.00_00_28_2             26.210             15.493        28
#> C1.0e+06_0.0e+00_00.00_00_35_1             27.029             15.523        35
#> C1.0e+06_0.0e+00_00.00_00_35_2             41.227             16.093        35
#>                                ConversionParameterB ConversionParameterC
#>                                               <Rle>                <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1            211703103           -146989184
#> C1.0e+05_1.0e+06_02.50_07_00_2            211703103           -146989184
#> C1.0e+05_1.0e+06_02.50_14_00_1            211703103           -146989183
#> C1.0e+05_1.0e+06_02.50_14_00_2            211703103           -146989183
#> C1.0e+05_1.0e+06_02.50_21_00              211703102           -146989182
#> ...                                             ...                  ...
#> C1.0e+06_0.0e+00_00.00_00_21_2            211702381           -146988181
#> C1.0e+06_0.0e+00_00.00_00_28_1            211702381           -146988180
#> C1.0e+06_0.0e+00_00.00_00_28_2            211702381           -146988180
#> C1.0e+06_0.0e+00_00.00_00_35_1            211702380           -146988179
#> C1.0e+06_0.0e+00_00.00_00_35_2            211702380           -146988179
#>                                TemperatureCompPpm SpaceChargeCompPpm  RawOvFtt
#>                                             <Rle>              <Rle> <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1             -14.44              -0.18  889335.3
#> C1.0e+05_1.0e+06_02.50_07_00_2             -14.44              -0.18  667088.6
#> C1.0e+05_1.0e+06_02.50_14_00_1             -14.44              -0.18  603502.7
#> C1.0e+05_1.0e+06_02.50_14_00_2             -14.44              -0.18  889290.8
#> C1.0e+05_1.0e+06_02.50_21_00               -14.44              -0.18  897244.2
#> ...                                           ...                ...       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2             -14.56              -1.76  247568.9
#> C1.0e+06_0.0e+00_00.00_00_28_1             -14.56              -1.76  187647.7
#> C1.0e+06_0.0e+00_00.00_00_28_2             -14.56              -1.76  130141.6
#> C1.0e+06_0.0e+00_00.00_00_35_1             -14.56              -1.76  191042.9
#> C1.0e+06_0.0e+00_00.00_00_35_2             -14.56              -1.76  268899.4
#>                                InjectionT0 NrCentroids SelectedMass1
#>                                  <numeric>   <integer>         <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1           0        3513        1211.7
#> C1.0e+05_1.0e+06_02.50_07_00_2           0        3053        1211.7
#> C1.0e+05_1.0e+06_02.50_14_00_1           0        3114        1211.7
#> C1.0e+05_1.0e+06_02.50_14_00_2           0        3743        1211.7
#> C1.0e+05_1.0e+06_02.50_21_00             0        3629        1211.7
#> ...                                    ...         ...           ...
#> C1.0e+06_0.0e+00_00.00_00_21_2      -0.054        4686       1211.72
#> C1.0e+06_0.0e+00_00.00_00_28_1       0.021        2640       1211.72
#> C1.0e+06_0.0e+00_00.00_00_28_2       0.045        2052       1211.72
#> C1.0e+06_0.0e+00_00.00_00_35_1       0.034        2123       1211.72
#> C1.0e+06_0.0e+00_00.00_00_35_2       0.036        2608       1211.72
#>                                Activation1 Energy1 SelectedMass2 Activation2
#>                                      <Rle>   <Rle>         <Rle>       <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1         ETD     2.5             0         CID
#> C1.0e+05_1.0e+06_02.50_07_00_2         ETD     2.5             0         CID
#> C1.0e+05_1.0e+06_02.50_14_00_1         ETD     2.5             0         CID
#> C1.0e+05_1.0e+06_02.50_14_00_2         ETD     2.5             0         CID
#> C1.0e+05_1.0e+06_02.50_21_00           ETD     2.5             0         CID
#> ...                                    ...     ...           ...         ...
#> C1.0e+06_0.0e+00_00.00_00_21_2         HCD      21            NA          NA
#> C1.0e+06_0.0e+00_00.00_00_28_1         HCD      28            NA          NA
#> C1.0e+06_0.0e+00_00.00_00_28_2         HCD      28            NA          NA
#> C1.0e+06_0.0e+00_00.00_00_35_1         HCD      35            NA          NA
#> C1.0e+06_0.0e+00_00.00_00_35_2         HCD      35            NA          NA
#>                                Energy2 EtdActivation CidActivation
#>                                  <Rle>         <Rle>         <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1       7           2.5             7
#> C1.0e+05_1.0e+06_02.50_07_00_2       7           2.5             7
#> C1.0e+05_1.0e+06_02.50_14_00_1      14           2.5            14
#> C1.0e+05_1.0e+06_02.50_14_00_2      14           2.5            14
#> C1.0e+05_1.0e+06_02.50_21_00        21           2.5            21
#> ...                                ...           ...           ...
#> C1.0e+06_0.0e+00_00.00_00_21_2      NA            NA            NA
#> C1.0e+06_0.0e+00_00.00_00_28_1      NA            NA            NA
#> C1.0e+06_0.0e+00_00.00_00_28_2      NA            NA            NA
#> C1.0e+06_0.0e+00_00.00_00_35_1      NA            NA            NA
#> C1.0e+06_0.0e+00_00.00_00_35_2      NA            NA            NA
#>                                HcdActivation Activation StartTime   EndTime
#>                                        <Rle>      <Rle> <numeric> <numeric>
#> C1.0e+05_1.0e+06_02.50_07_00_1            NA      ETcid      6.01       6.5
#> C1.0e+05_1.0e+06_02.50_07_00_2            NA      ETcid      6.01       6.5
#> C1.0e+05_1.0e+06_02.50_14_00_1            NA      ETcid      7.01       7.5
#> C1.0e+05_1.0e+06_02.50_14_00_2            NA      ETcid      7.01       7.5
#> C1.0e+05_1.0e+06_02.50_21_00              NA      ETcid      8.01       8.5
#> ...                                      ...        ...       ...       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2            21        HCD     79.01      79.5
#> C1.0e+06_0.0e+00_00.00_00_28_1            28        HCD     80.01      80.5
#> C1.0e+06_0.0e+00_00.00_00_28_2            28        HCD     80.01      80.5
#> C1.0e+06_0.0e+00_00.00_00_35_1            35        HCD     81.01      81.5
#> C1.0e+06_0.0e+00_00.00_00_35_2            35        HCD     81.01      81.5
#>                                ActivationType    Ce SteppedCe SteppedCeValue
#>                                         <Rle> <Rle>     <Rle>          <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1            ETD    NA                       NA
#> C1.0e+05_1.0e+06_02.50_07_00_2            ETD    NA                       NA
#> C1.0e+05_1.0e+06_02.50_14_00_1            ETD    NA                       NA
#> C1.0e+05_1.0e+06_02.50_14_00_2            ETD    NA                       NA
#> C1.0e+05_1.0e+06_02.50_21_00              ETD    NA                       NA
#> ...                                       ...   ...       ...            ...
#> C1.0e+06_0.0e+00_00.00_00_21_2            HCD    25     False              5
#> C1.0e+06_0.0e+00_00.00_00_28_1            HCD    25     False              5
#> C1.0e+06_0.0e+00_00.00_00_28_2            HCD    25     False              5
#> C1.0e+06_0.0e+00_00.00_00_35_1            HCD    25     False              5
#> C1.0e+06_0.0e+00_00.00_00_35_2            HCD    25     False              5
#>                                MultistageActivation Ethcd
#>                                               <Rle> <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1                      FALSE
#> C1.0e+05_1.0e+06_02.50_07_00_2                      FALSE
#> C1.0e+05_1.0e+06_02.50_14_00_1                      FALSE
#> C1.0e+05_1.0e+06_02.50_14_00_2                      FALSE
#> C1.0e+05_1.0e+06_02.50_21_00                        FALSE
#> ...                                             ...   ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                      FALSE
#> C1.0e+06_0.0e+00_00.00_00_28_1                      FALSE
#> C1.0e+06_0.0e+00_00.00_00_28_2                      FALSE
#> C1.0e+06_0.0e+00_00.00_00_35_1                      FALSE
#> C1.0e+06_0.0e+00_00.00_00_35_2                      FALSE
#>                                UseCalibratedEtdParameters EtdReactionTime
#>                                                     <Rle>           <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1                      False             2.5
#> C1.0e+05_1.0e+06_02.50_07_00_2                      False             2.5
#> C1.0e+05_1.0e+06_02.50_14_00_1                      False             2.5
#> C1.0e+05_1.0e+06_02.50_14_00_2                      False             2.5
#> C1.0e+05_1.0e+06_02.50_21_00                        False             2.5
#> ...                                                   ...             ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                                         NA
#> C1.0e+06_0.0e+00_00.00_00_28_1                                         NA
#> C1.0e+06_0.0e+00_00.00_00_28_2                                         NA
#> C1.0e+06_0.0e+00_00.00_00_35_1                                         NA
#> C1.0e+06_0.0e+00_00.00_00_35_2                                         NA
#>                                EtdReagentTarget EtdSupplementalActivation
#>                                           <Rle>                     <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1          1000000                      True
#> C1.0e+05_1.0e+06_02.50_07_00_2          1000000                      True
#> C1.0e+05_1.0e+06_02.50_14_00_1          1000000                      True
#> C1.0e+05_1.0e+06_02.50_14_00_2          1000000                      True
#> C1.0e+05_1.0e+06_02.50_21_00            1000000                      True
#> ...                                         ...                       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2               NA                          
#> C1.0e+06_0.0e+00_00.00_00_28_1               NA                          
#> C1.0e+06_0.0e+00_00.00_00_28_2               NA                          
#> C1.0e+06_0.0e+00_00.00_00_35_1               NA                          
#> C1.0e+06_0.0e+00_00.00_00_35_2               NA                          
#>                                SupplementalActivationCe AgcTarget Activationq
#>                                                   <Rle>     <Rle>       <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1                        7    100000          NA
#> C1.0e+05_1.0e+06_02.50_07_00_2                        7    100000          NA
#> C1.0e+05_1.0e+06_02.50_14_00_1                       14    100000          NA
#> C1.0e+05_1.0e+06_02.50_14_00_2                       14    100000          NA
#> C1.0e+05_1.0e+06_02.50_21_00                         21    100000          NA
#> ...                                                 ...       ...         ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                       NA   1000000          NA
#> C1.0e+06_0.0e+00_00.00_00_28_1                       NA   1000000          NA
#> C1.0e+06_0.0e+00_00.00_00_28_2                       NA   1000000          NA
#> C1.0e+06_0.0e+00_00.00_00_35_1                       NA   1000000          NA
#> C1.0e+06_0.0e+00_00.00_00_35_2                       NA   1000000          NA
#>                                        TargetedMassList     Mz Charge    Sample
#>                                             <character>  <Rle>  <Rle> <integer>
#> C1.0e+05_1.0e+06_02.50_07_00_1 (mz=1211.7012 z=1 name=) 1211.7     14         1
#> C1.0e+05_1.0e+06_02.50_07_00_2 (mz=1211.7012 z=1 name=) 1211.7     14         1
#> C1.0e+05_1.0e+06_02.50_14_00_1 (mz=1211.7014 z=1 name=) 1211.7     14         2
#> C1.0e+05_1.0e+06_02.50_14_00_2 (mz=1211.7014 z=1 name=) 1211.7     14         2
#> C1.0e+05_1.0e+06_02.50_21_00   (mz=1211.7016 z=1 name=) 1211.7     14         3
#> ...                                                 ...    ...    ...       ...
#> C1.0e+06_0.0e+00_00.00_00_21_2 (mz=1211.7158 z=1 name=) 1211.7     14       220
#> C1.0e+06_0.0e+00_00.00_00_28_1  (mz=1211.716 z=1 name=) 1211.7     14       221
#> C1.0e+06_0.0e+00_00.00_00_28_2  (mz=1211.716 z=1 name=) 1211.7     14       221
#> C1.0e+06_0.0e+00_00.00_00_35_1 (mz=1211.7162 z=1 name=) 1211.7     14       222
#> C1.0e+06_0.0e+00_00.00_00_35_2 (mz=1211.7162 z=1 name=) 1211.7     14       222
#>                                MedianIonInjectionTimeMs
#>                                                   <Rle>
#> C1.0e+05_1.0e+06_02.50_07_00_1                    4.837
#> C1.0e+05_1.0e+06_02.50_07_00_2                    4.837
#> C1.0e+05_1.0e+06_02.50_14_00_1                    4.837
#> C1.0e+05_1.0e+06_02.50_14_00_2                    4.837
#> C1.0e+05_1.0e+06_02.50_21_00                      4.837
#> ...                                                 ...
#> C1.0e+06_0.0e+00_00.00_00_21_2                    43.14
#> C1.0e+06_0.0e+00_00.00_00_28_1                    43.14
#> C1.0e+06_0.0e+00_00.00_00_28_2                    43.14
#> C1.0e+06_0.0e+00_00.00_00_35_1                    43.14
#> C1.0e+06_0.0e+00_00.00_00_35_2                    43.14

head(assayData(tds))

#> 6 x 351 sparse Matrix of class "dgCMatrix"

#>    [[ suppressing 351 column names ‘C1.0e+05_1.0e+06_02.50_07_00_1’, ‘C1.0e+05_1.0e+06_02.50_07_00_2’, ‘C1.0e+05_1.0e+06_02.50_14_00_1’ ... ]]

#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                                                               
#> a1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . . .
#>                                         
#> a1   . . . . . . . . . . . . . . . . . .
#> b1   . . . . . . . . . . . . . . . . . .
#> z1   . . . . . . . . . . . . . . . . . .
#> zpH1 . . . . . . . . . . . . . . . . . .
#> c1   . . . . . . . . . . . . . . . . . .
#> y1   . . . . . . . . . . . . . . . . . .

# Accessing colData
tds$Mz

#> numeric-Rle of length 351 with 1 run
#>   Lengths:    351
#>   Values : 1211.7

tds$FilterString

#> NULL

# Subsetting

# First 100 fragments
tds[1:100]

#> TopDownSet object (0.31 Mb)
#> - - - Protein data - - -
#> Amino acid sequence (153): GLSDGEWQQVLNVWGKVEADIAGHGQ...GAMTKALELFRNDIAAKYKELGFQG 
#> Mass : 16964.96
#> Modifications (3): Carbamidomethyl, Acetyl, Met-loss
#> - - - Fragment data - - -
#> Number of theoretical fragments: 100 
#> Theoretical fragment types (12): a, a*, b, b*, c, ..., x*, y, y*, z, z*
#> Theoretical mass range: [72.04;1142.51]
#> - - - Condition data - - -
#> Number of conditions: 222 
#> Number of scans: 351 
#> Condition variables (57): Scan, PeaksCount, ..., Sample, MedianIonInjectionTimeMs
#> - - - Intensity data - - -
#> Size of array: 100x351 (3.30% != 0)
#> Number of matched fragments: 1157 
#> Intensity range: [280.27;214724.06]
#> - - - Processing information - - -
#> [2017-12-28 16:44:52] 14901 fragments [1949;351] matched (tolerance: 25 ppm).
#> [2017-12-28 16:44:52] Condition names updated based on: Mz, AgcTarget, EtdReagentTarget, EtdActivation, CidActivation, HcdActivation. Order of conditions changed. 222 conditions.
#> [2017-12-28 16:44:52] Recalculate median injection time based on: Mz, AgcTarget.
#> [2017-12-28 16:52:26] Subsetted 14901 fragments [1949;351] to 1157 fragments [100;351]. 

# All c fragments
tds["c"]

#> TopDownSet object (0.30 Mb)
#> - - - Protein data - - -
#> Amino acid sequence (153): GLSDGEWQQVLNVWGKVEADIAGHGQ...GAMTKALELFRNDIAAKYKELGFQG 
#> Mass : 16964.96
#> Modifications (3): Carbamidomethyl, Acetyl, Met-loss
#> - - - Fragment data - - -
#> Number of theoretical fragments: 152 
#> Theoretical fragment types (1): c
#> Theoretical mass range: [117.07;16925.98]
#> - - - Condition data - - -
#> Number of conditions: 222 
#> Number of scans: 351 
#> Condition variables (57): Scan, PeaksCount, ..., Sample, MedianIonInjectionTimeMs
#> - - - Intensity data - - -
#> Size of array: 152x351 (0.68% != 0)
#> Number of matched fragments: 364 
#> Intensity range: [432.91;664294.75]
#> - - - Processing information - - -
#> [2017-12-28 16:44:52] 14901 fragments [1949;351] matched (tolerance: 25 ppm).
#> [2017-12-28 16:44:52] Condition names updated based on: Mz, AgcTarget, EtdReagentTarget, EtdActivation, CidActivation, HcdActivation. Order of conditions changed. 222 conditions.
#> [2017-12-28 16:44:52] Recalculate median injection time based on: Mz, AgcTarget.
#> [2017-12-28 16:52:26] Subsetted 14901 fragments [1949;351] to 364 fragments [152;351]. 

# Just c 152
tds["c152"]

#> TopDownSet object (0.28 Mb)
#> - - - Protein data - - -
#> Amino acid sequence (153): GLSDGEWQQVLNVWGKVEADIAGHGQ...GAMTKALELFRNDIAAKYKELGFQG 
#> Mass : 16964.96
#> Modifications (3): Carbamidomethyl, Acetyl, Met-loss
#> - - - Fragment data - - -
#> Number of theoretical fragments: 1 
#> Theoretical fragment types (1): c
#> Theoretical mass range: [16925.98;16925.98]
#> - - - Condition data - - -
#> Number of conditions: 222 
#> Number of scans: 351 
#> Condition variables (57): Scan, PeaksCount, ..., Sample, MedianIonInjectionTimeMs
#> - - - Intensity data - - -
#> Size of array: 1x351 (38.46% != 0)
#> Number of matched fragments: 135 
#> Intensity range: [979.57;664294.75]
#> - - - Processing information - - -
#> [2017-12-28 16:44:52] 14901 fragments [1949;351] matched (tolerance: 25 ppm).
#> [2017-12-28 16:44:52] Condition names updated based on: Mz, AgcTarget, EtdReagentTarget, EtdActivation, CidActivation, HcdActivation. Order of conditions changed. 222 conditions.
#> [2017-12-28 16:44:52] Recalculate median injection time based on: Mz, AgcTarget.
#> [2017-12-28 16:52:26] Subsetted 14901 fragments [1949;351] to 135 fragments [1;351]. 

# Condition 1 to 10
tds[, 1:10]

#> TopDownSet object (0.35 Mb)
#> - - - Protein data - - -
#> Amino acid sequence (153): GLSDGEWQQVLNVWGKVEADIAGHGQ...GAMTKALELFRNDIAAKYKELGFQG 
#> Mass : 16964.96
#> Modifications (3): Carbamidomethyl, Acetyl, Met-loss
#> - - - Fragment data - - -
#> Number of theoretical fragments: 1949 
#> Theoretical fragment types (12): a, a*, b, b*, c, ..., x*, y, y*, z, z*
#> Theoretical mass range: [72.04;16925.98]
#> - - - Condition data - - -
#> Number of conditions: 6 
#> Number of scans: 10 
#> Condition variables (57): Scan, PeaksCount, ..., Sample, MedianIonInjectionTimeMs
#> - - - Intensity data - - -
#> Size of array: 1949x10 (2.78% != 0)
#> Number of matched fragments: 542 
#> Intensity range: [1843.09;669750.88]
#> - - - Processing information - - -
#> [2017-12-28 16:44:52] 14901 fragments [1949;351] matched (tolerance: 25 ppm).
#> [2017-12-28 16:44:52] Condition names updated based on: Mz, AgcTarget, EtdReagentTarget, EtdActivation, CidActivation, HcdActivation. Order of conditions changed. 222 conditions.
#> [2017-12-28 16:44:52] Recalculate median injection time based on: Mz, AgcTarget.
#> [2017-12-28 16:52:26] Subsetted 14901 fragments [1949;351] to 542 fragments [1949;10].