The TopDownSet class

The TopDownSet class is a container for a whole top-down proteomics experiment.

# S4 method for TopDownSet
aggregate(x, by = x$Sample, ...)

# S4 method for TopDownSet
filterCv(object, threshold, by = object$Sample, ...)

# S4 method for TopDownSet
filterInjectionTime(object, maxDeviation = log2(3),
  keepTopN = 2, by = object$Sample, ...)

# S4 method for TopDownSet
filterIntensity(object, threshold, relative = TRUE,
  ...)

# S4 method for TopDownSet
filterNonReplicatedFragments(object, minN = 2,
  by = object$Sample, ...)

# S4 method for TopDownSet
normalize(object, method = "TIC", ...)

# S4 method for TopDownSet
show(object)

# S4 method for TopDownSet
summary(object, what = c("conditions", "fragments"),
  ...)

Arguments

x, object
TopDownSet

by
list, grouping variable, in general it refers to technical

threshold
double, threshold variable.

maxDeviation
double, maximal allowed deviation in the log2 injection time in ms in comparison to the median ion injection time.

keepTopN
integer, how many technical replicates should be kept?

relative
logical, if relative is TRUE all fragments with intensity below threshold * max(intensity) per fragment are removed, otherwise all fragments below threshold are removed.

minN
numeric, if less than minN of a fragment are found across technical replicates it is removed.

method
character, normalisation method, currently just "TIC" for Total Ion Current normalisation of the scans/conditions (column-wise normalisation) is supported.

what
character, specifies whether "conditions" (columns; default) or "fragments" (rows) should be summarized.

…
arguments passed to internal/other methods. replicates (that's why the default is the "Sample" column in colData).

Details

See vignette("analysis", package="topdownr") for a detailed example how to work with TopDownSet objects.

Methods (by generic)

aggregate: Aggregate conditions/runs.

Aggregates conditions/runs (columns) in an TopDownSet object by a user-given value (default is the "Sample" column of colData which has the same value for technical replicates). It combines intensity values and numeric metadata of the grouped conditions/runs (columns) by mean and returns a reduced TopDownSet object.

filterCv: Filter by CV.

Filtering is done by coefficient of variation across technical replicates (defined by the by argument). All fragments below a given threshold are removed. The threshold is the maximal allowed CV in percent (sd/mean * 100 &lt; threshold).

filterInjectionTime: Filter by ion injection time.

Filtering is done by maximal allowed deviation and just the technical keepTopN replicates with the lowest deviation from the median ion injection time are kept.

filterIntensity: Filter by intensity.

Filtering is done by removing all fragments that are below a given (absolute/relative) intensity threshold. threshold.

filterNonReplicatedFragments: Filter by non-replicated fragments.

Filtering is done by removing all fragments that don't replicate across technical replicates.

normalize: Normalise.

Applies Total Ion Current normalisation to a TopDownSet. The normalisation ist done per scans/conditions (column-wise normalisation).

summary: Summary statistics.

Returns a matrix with some statistics: number of fragments, total/min/first quartile/median/mean/third quartile/maximum of intensity values.

Slots

rowViews

FragmentViews, information about fragments (name, type, sequence, mass, charge), see FragmentViews for details.

colData

S4Vectors::DataFrame, information about the MS2 experiments and the fragmentation conditions.

assay

Matrix::dgCMatrix, intensity values of the fragments. The rows correspond to the fragments and the columns to the condition/run. It just stores values that are different from zero.

files

character, files that were imported.

tolerance

double, tolerance in ppm that were used for matching the experimental mz values to the theoretical fragments.

processing

character, log messages.

Coercion

`as(object, "MSnSet"): Coerce an TopDownSet object into an MSnbase::MSnSet object.

`as(object, "NCBSet"): Coerce an TopDownSet object into an NCBSet object.

See also

FragmentViews for the row view interface.

Matrix::dgCMatrix for technical details about the intensity storage.