Investigation of Fragmentation Conditions in Top-Down Proteomics

The topdownr package allows automatic and systemic investigation of fragment conditions. It creates Thermo Orbitrap Fusion Lumos method files to test hundreds of fragmentation conditions. Additionally it provides functions to analyse and process the generated MS data and determine the best conditions to maximise overall fragment coverage.

Details

The usage of the topdownr package is demonstrated in the following vignettes:

Generate .meth files prior data acquisition for the Thermo Orbitrap Fusion Lumos MS devise: vignette("data-generation", package="topdownr").

How to analyse top-down fragmenation data: vignette("analysis", package="topdownr")

References

https://github.com/sgibb/topdownr/

See also

Useful links:

https://github.com/sgibb/topdownr/

Report bugs at https://github.com/sgibb/topdownr/issues/